tests/artifacts/summarize-url-content/requirements.txt
any-agent[all]==0.22.0
python-dotenv
fire
beautifulsoup4
requests